agent/requirements.txt
fastapi
uvicorn
requests
google-cloud-aiplatform
python-multipart
python-dotenv